Tabs™ are here
Tabs are now in the modal. Added a lot of things in the surge.css and surge.js to get tabs working inside of the modal. They work, but will need to be formatted to look nicer, will take care of that tomorrow. also went ahead and commented some things for the new guys who wanna read the code.